requirements1.txt
beautifulsoup4
biorefineries
biosteam
fastapi
markdownify
matplotlib
nest_asyncio
numpy
numpy_financial
openai
pydantic
pydash
pypdf
pyppeteer
python-dotenv
transitions
uvicorn
gdrive
gspread
gspread_dataframe